Override the getEula() method at the SnipeModel level
